integrations/llm-agent-frameworks/modaic/pyproject.toml
[project]
name = "modaic-recipe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "modaic>=0.4.1",
    "weaviate-client>=4.18.1",
]
